리스트수정

+++ src/main/webapp/space/sub01_view.jsp
@@ -134,8 +134,15 @@ function deletePost(){
 	<div class="row mb-3">
 		<div class="col d-flex justify-content-end">
 			<!-- 각종 버튼 부분 -->
+			 <%
+            if(session.getAttribute("user_id") != null
+            && session.getAttribute("user_id").toString().equals(dto.getId())){
+            %>   
 			<button type="button" class="btn btn-primary btn-sm" onclick="location.href='Edit.jsp?idx=<%=dto.getIdx()%>&tname=<%=tname%>';">수정하기</button>
 			<button type="button" class="btn btn-success btn-sm" onclick="deletePost();">삭제하기</button>	
+			<%
+            }
+			%>
 			<button type="button" class="btn btn-warning btn-sm" 
 				onclick="location.href='ListSkin.jsp';">리스트보기</button>
 		</div>
